# openapi/lis-internal-v1.yaml
openapi: 3.1.0
info:
  title: LIS Core API
  summary: Starter modular monolith for a laboratory information system
  description: Internal API with working persistence for auth, master data, orders,
    specimen lifecycle, task orchestration, observations, reports, audit, provenance,
    device registry, HL7 v2 integrations, device gateway, autoverification, ASTM-style
    drivers, and a read/search FHIR facade.
  version: 0.7.0
paths:
...
  /fhir/R4/Patient:
    get:
      tags:
      - fhir
      summary: Search Patients
      operationId: search_patients_fhir_R4_Patient_get
      security:
      - HTTPBearer: []
      parameters:
      - name: identifier
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Identifier
      - name: family
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Family
      - name: _id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: ' Id'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Search Patients Fhir R4 Patient Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /fhir/R4/Patient/{patient_id}:
    get:
      tags:
      - fhir
      summary: Read Patient
      operationId: read_patient_fhir_R4_Patient__patient_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Patient Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Read Patient Fhir R4 Patient  Patient Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /fhir/R4/ServiceRequest:
    get:
      tags:
      - fhir
      summary: Search Service Requests
      operationId: search_service_requests_fhir_R4_ServiceRequest_get
      security:
      - HTTPBearer: []
      parameters:
      - name: patient
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Patient
      - name: identifier
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Identifier
      - name: requisition
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Requisition
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: _id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: ' Id'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Search Service Requests Fhir R4 Servicerequest Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /fhir/R4/ServiceRequest/{service_request_id}:
    get:
      tags:
      - fhir
      summary: Read Service Request
      operationId: read_service_request_fhir_R4_ServiceRequest__service_request_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: service_request_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Service Request Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Read Service Request Fhir R4 Servicerequest  Service
                  Request Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /fhir/R4/Specimen:
    get:
      tags:
      - fhir
      summary: Search Specimens
      operationId: search_specimens_fhir_R4_Specimen_get
      security:
      - HTTPBearer: []
      parameters:
      - name: patient
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Patient
      - name: identifier
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Identifier
      - name: accession
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Accession
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: _id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: ' Id'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Search Specimens Fhir R4 Specimen Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /fhir/R4/Specimen/{specimen_id}:
    get:
      tags:
      - fhir
      summary: Read Specimen
      operationId: read_specimen_fhir_R4_Specimen__specimen_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: specimen_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Specimen Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Read Specimen Fhir R4 Specimen  Specimen Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /fhir/R4/Task:
    get:
      tags:
      - fhir
      summary: Search Tasks
      operationId: search_tasks_fhir_R4_Task_get
      security:
      - HTTPBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: focus
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Focus
      - name: code
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Code
      - name: patient
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Patient
      - name: _id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: ' Id'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Search Tasks Fhir R4 Task Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /fhir/R4/Task/{task_id}:
    get:
      tags:
      - fhir
      summary: Read Task
      operationId: read_task_fhir_R4_Task__task_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Task Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Read Task Fhir R4 Task  Task Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /fhir/R4/Observation:
    get:
      tags:
      - fhir
      summary: Search Observations
      operationId: search_observations_fhir_R4_Observation_get
      security:
      - HTTPBearer: []
      parameters:
      - name: patient
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Patient
      - name: specimen
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Specimen
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: code
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Code
      - name: based-on
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Based-On
      - name: _id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: ' Id'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Search Observations Fhir R4 Observation Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /fhir/R4/Observation/{observation_id}:
    get:
      tags:
      - fhir
      summary: Read Observation
      operationId: read_observation_fhir_R4_Observation__observation_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: observation_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Observation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Read Observation Fhir R4 Observation  Observation
                  Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /fhir/R4/DiagnosticReport:
    get:
      tags:
      - fhir
      summary: Search Diagnostic Reports
      operationId: search_diagnostic_reports_fhir_R4_DiagnosticReport_get
      security:
      - HTTPBearer: []
      parameters:
      - name: patient
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Patient
      - name: based-on
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Based-On
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: _id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: ' Id'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Search Diagnostic Reports Fhir R4 Diagnosticreport
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /fhir/R4/DiagnosticReport/{report_id}:
    get:
      tags:
      - fhir
      summary: Read Diagnostic Report
      operationId: read_diagnostic_report_fhir_R4_DiagnosticReport__report_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: report_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Report Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Read Diagnostic Report Fhir R4 Diagnosticreport  Report
                  Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /fhir/R4/AuditEvent:
    get:
      tags:
      - fhir
      summary: Search Audit Events
      operationId: search_audit_events_fhir_R4_AuditEvent_get
      security:
      - HTTPBearer: []
      parameters:
      - name: entity
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Entity
      - name: _id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: ' Id'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Search Audit Events Fhir R4 Auditevent Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /fhir/R4/AuditEvent/{audit_event_id}:
    get:
      tags:
      - fhir
      summary: Read Audit Event
      operationId: read_audit_event_fhir_R4_AuditEvent__audit_event_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: audit_event_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Audit Event Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Read Audit Event Fhir R4 Auditevent  Audit Event Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /fhir/R4/Provenance:
    get:
      tags:
      - fhir
      summary: Search Provenance
      operationId: search_provenance_fhir_R4_Provenance_get
      security:
      - HTTPBearer: []
      parameters:
      - name: target
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Target
      - name: _id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: ' Id'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Search Provenance Fhir R4 Provenance Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /fhir/R4/Provenance/{provenance_id}:
    get:
      tags:
      - fhir
      summary: Read Provenance
      operationId: read_provenance_fhir_R4_Provenance__provenance_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: provenance_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Provenance Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Read Provenance Fhir R4 Provenance  Provenance Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/provenance:
    get:
      tags:
      - audit
      summary: Search Provenance
      operationId: search_provenance_api_v1_provenance_get
      security:
      - HTTPBearer: []
      parameters:
      - name: target_resource_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Target Resource Type
      - name: target_resource_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Target Resource Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    $ref: '#/components/schemas/ProvenanceRecordSummary'
                title: Response Search Provenance Api V1 Provenance Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ProvenanceRecordSummary:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        target_resource_type:
          type: string
          title: Target Resource Type
        target_resource_id:
          type: string
          format: uuid
          title: Target Resource Id
        activity_code:
          type: string
          title: Activity Code
        based_on_order_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Based On Order Id
        based_on_order_item_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Based On Order Item Id
        specimen_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Specimen Id
        observation_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Observation Id
        report_version_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Report Version Id
        device_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Device Id
        agent_user_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Agent User Id
        agent_practitioner_role_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Agent Practitioner Role Id
        recorded_at:
          type: string
          format: date-time
          title: Recorded At
        inputs:
          additionalProperties: true
          type: object
          title: Inputs
        signature:
          additionalProperties: true
          type: object
          title: Signature
      additionalProperties: false
      type: object
      required:
      - id
      - target_resource_type
      - target_resource_id
      - activity_code
      - recorded_at
      title: ProvenanceRecordSummary
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer